Made new package for rendering related stuff
